feat: address ESLint's `no-unresolved` warning
This adds the `main` property for backwards-compatibility, needed e.g.
when running ESLint.

The change recapitulates what
https://github.com/octokit/rest.js/pull/413/commits/aad55f47f1e6348fa3d246697d7503985eff4c06
already did but which was unfortunately reverted in the same PR by
https://github.com/octokit/rest.js/pull/413/commits/1b6e5829faaaefb9e4133685f719a56234a304b1
(probably under the incorrect assumption that `exports.default` would be
enough as a fall-back).

This fixes https://github.com/octokit/rest.js/issues/446.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>